requirements.txt
Pillow
dearpygui
requests
Jinja2
bs4
lxml
pyinstaller